% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{asy_test}
\alias{asy_test}
\title{Asymptotic test}
\usage{
asy_test(data, k, hyper = NULL, kernel_choice = "Linear")
}
\arguments{
\item{data}{Data containing observed sample X, sample Y and covariate Z}

\item{k}{Number of nearest neighbors used}

\item{hyper}{Hyperparameter used in kernel function}

\item{kernel_choice}{Choice of kernel function}
}
\value{
Normalized test statistic and p-value
}
\description{
Asymptotic test
}
