% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{asy_var}
\alias{asy_var}
\title{Compute asymptotic conditional variance estimate}
\usage{
asy_var(k, X, Y, nn_ind, l, kernel_choice = "Linear", num_derandom = 1)
}
\arguments{
\item{k}{Number of nearest neighbors}

\item{X}{Observed sample X}

\item{Y}{Observed sample Y}

\item{nn_ind}{earest neighbor matrix of size n by k; the first column is the i-th unit value}

\item{l}{The hyperparameter}

\item{kernel_choice}{Choice of kernel function, the default is linear kernel}

\item{num_derandom}{Number of derandomized sample used}
}
\value{
asymptotic variance estimate
}
\description{
Compute asymptotic conditional variance estimate
}
