% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{bootstrap_SKCE_classification}
\alias{bootstrap_SKCE_classification}
\title{Bootstrap for SKCE classification calibration}
\usage{
bootstrap_SKCE_classification(predictor, gamma = 1, B = 500)
}
\arguments{
\item{predictor}{Predictor matrix}

\item{gamma}{Hyperparameter in Laplace kernel}

\item{B}{Number of bootstrap}
}
\value{
A vector of bootstrap sample of length B
}
\description{
Bootstrap for SKCE classification calibration
}
