% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{bootstrap_SKCE_regression}
\alias{bootstrap_SKCE_regression}
\title{Bootstrap for SKCE regression}
\usage{
bootstrap_SKCE_regression(pseduo_data, gamma = 1, B = 500)
}
\arguments{
\item{pseduo_data}{Same for SKCE_regression}

\item{gamma}{Same for SKCE_regression}

\item{B}{Number of bootstrap}
}
\value{
A vector of bootstrap sample under null
}
\description{
Bootstrap for SKCE regression
}
