% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{compute_ECE}
\alias{compute_ECE}
\title{Compute the expected calibration error}
\usage{
compute_ECE(predicted_probs, true_labels, n_bins = 10)
}
\arguments{
\item{predicted_probs}{A vector of predicted probabilities for the positive class.}

\item{true_labels}{A vector of true labels (0 or 1).}

\item{n_bins}{Number of bins to use.}
}
\value{
ECE value
}
\description{
Compute the expected calibration error
}
