% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{data_reliability_diagram}
\alias{data_reliability_diagram}
\title{Data preparation for plotting reliability diagram}
\usage{
data_reliability_diagram(predicted_probs, true_labels, n_bins = 10)
}
\arguments{
\item{predicted_probs}{Predicted probability vector}

\item{true_labels}{True observation label}

\item{n_bins}{Number of bins used}
}
\value{
A dataframe containing necessary ingredients for reliability diagram
}
\description{
Data preparation for plotting reliability diagram
}
