% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{derandomized_test}
\alias{derandomized_test}
\title{Derandomized test}
\usage{
derandomized_test(
  data,
  k,
  hyper = NULL,
  kernel_choice = "Linear",
  num_derandom = 300,
  resampling_dist = "Gaussian",
  resamp_hyper = list(mean = 0, sd = 1)
)
}
\arguments{
\item{data}{Same as df_test}

\item{k}{Same as df_test}

\item{hyper}{Same as df_test}

\item{kernel_choice}{Same as df_test}

\item{num_derandom}{Number of derandomized sample used}

\item{resampling_dist}{Same as df_test}

\item{resamp_hyper}{Same as df_test}
}
\value{
Nomalized test statistic and p-value
}
\description{
Derandomized test
}
