% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{df_test}
\alias{df_test}
\title{Distribution-free finite-sample valid test}
\usage{
df_test(
  data,
  k,
  hyper = NULL,
  kernel_choice = "Linear",
  M = 300,
  resampling_dist = "Gaussian",
  resamp_hyper = list(mean = 0, sd = 1)
)
}
\arguments{
\item{data}{Data containing observed sample Y and covariate Z}

\item{k}{Number of nearest neighbors}

\item{hyper}{Hyperparameter used in kernel function}

\item{kernel_choice}{Choice of kernel}

\item{M}{Number of resamples}

\item{resampling_dist}{Resampling distribution}

\item{resamp_hyper}{Hyperparameter in the resampling distribution}
}
\value{
P-value
}
\description{
Distribution-free finite-sample valid test
}
