% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{isotonic.calibration}
\alias{isotonic.calibration}
\title{Return calibration using Isotonic regression (fit the test point x0 with the model)}
\usage{
isotonic.calibration(iso, x0)
}
\arguments{
\item{iso}{Isotonic regression obtained from isoreg function}

\item{x0}{Test point, predicted probability vector}
}
\value{
Calibrated predicted probability
}
\description{
Return calibration using Isotonic regression (fit the test point x0 with the model)
}
