% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{isotonic_calibration}
\alias{isotonic_calibration}
\title{Isotonic calibration function}
\usage{
isotonic_calibration(y, p, p_test)
}
\arguments{
\item{y}{Training observation for Isotonic regression (a subset of test data)}

\item{p}{Prediction probability obtained from some trained model}

\item{p_test}{Prediction proabability on the rest of data}
}
\value{
Calibrated predicted probability
}
\description{
Isotonic calibration function
}
