% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{KCSD_test}
\alias{KCSD_test}
\title{KCSD test (for conditional goodness-of-fit test)}
\usage{
KCSD_test(data, mu, sigma, M = 200)
}
\arguments{
\item{data}{Data consisting observed outcome Y and conditional covariate X}

\item{mu}{Mean function x}

\item{sigma}{Standard deviation function x}

\item{M}{Number of bootstrap samples}
}
\value{
Bootstrap-based p-value
}
\description{
KCSD test (for conditional goodness-of-fit test)
}
