% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{kernel}
\alias{kernel}
\title{Kernel function}
\usage{
kernel(X, Y, l = NULL, kernel_choice = "Linear")
}
\arguments{
\item{X}{First argument}

\item{Y}{Second argument}

\item{l}{The hyperparameter}

\item{kernel_choice}{Choice of kernel function}
}
\value{
Kernel value
}
\description{
Kernel function
}
