% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{kernel_multivariate}
\alias{kernel_multivariate}
\title{Multivariate kernel function}
\usage{
kernel_multivariate(X, Y, l = NULL, kernel_choice = "Linear")
}
\arguments{
\item{X}{First argument with dimension greater or equal than 2}

\item{Y}{Second argument with dimension greater or equal than 2}

\item{l}{Bandwidth}

\item{kernel_choice}{Choice of kernel}
}
\value{
Kernel value
}
\description{
Multivariate kernel function
}
