% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{kernel_Ustat}
\alias{kernel_Ustat}
\title{Kernel function in the U-statistic}
\usage{
kernel_Ustat(x1, x2, y1, y2, lx, ly, mu, sigma)
}
\arguments{
\item{x1}{First argument of kernel function in X space}

\item{x2}{Second argument of kernel function in X space}

\item{y1}{First argument of kernel function in Y space}

\item{y2}{Second argument of kernel function in Y space}

\item{lx}{Bandwidth chosen for kernel function in X space}

\item{ly}{Bandwidth chosen for kernel function in Y space}

\item{mu}{Mean function}

\item{sigma}{Standard deviation function}
}
\value{
Kernel value in the U-stat
}
\description{
Kernel function in the U-statistic
}
