% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{score}
\alias{score}
\title{Score function}
\usage{
score(x, y, mu, sigma)
}
\arguments{
\item{x}{Conditioning argument}

\item{y}{Gaussian argument}

\item{mu}{Mean function}

\item{sigma}{Standard deviation function}
}
\value{
Score value evaluated at y|x
}
\description{
Score function
}
