% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{SKCE_classification}
\alias{SKCE_classification}
\title{SKCE estimator for classification calibration}
\usage{
SKCE_classification(
  predictor,
  response,
  weight_mat = matrix(1, nrow(predictor), nrow(predictor)),
  gamma = 2
)
}
\arguments{
\item{predictor}{Predictor matrix}

\item{response}{Response vector}

\item{weight_mat}{Weight matrix of size n by n}

\item{gamma}{Hyperparameter in Laplace dot kernel}
}
\value{
SKCE estimator
}
\description{
SKCE estimator for classification calibration
}
