% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{SKCE_classification_test}
\alias{SKCE_classification_test}
\title{SKCE test (for classification model calibration test)}
\usage{
SKCE_classification_test(data, B = 300)
}
\arguments{
\item{data}{Data consisting predictor and response matrix}

\item{B}{Number of bootstrap}
}
\value{
P-value
}
\description{
SKCE test (for classification model calibration test)
}
