% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{SKCE_regression}
\alias{SKCE_regression}
\title{SKCE for regression calibration}
\usage{
SKCE_regression(
  pseduo_data,
  weight_mat = matrix(1, nrow(pseduo_data), nrow(pseduo_data)),
  gamma = 1
)
}
\arguments{
\item{pseduo_data}{Containing observed data and sufficient statistics}

\item{weight_mat}{Weight matrix used in constructing test statistic}

\item{gamma}{Hyperparameter in Laplace kernel}
}
\value{
SKCE value
}
\description{
SKCE for regression calibration
}
