% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{SKCE_regression_test}
\alias{SKCE_regression_test}
\title{SKCE test for regression calibration}
\usage{
SKCE_regression_test(data, B = 300)
}
\arguments{
\item{data}{Pseduo data}

\item{B}{Number of bootstrap}
}
\value{
P-value
}
\description{
SKCE test for regression calibration
}
