% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{test_stat}
\alias{test_stat}
\title{Define test statistic}
\usage{
test_stat(k, X, Y, nn_ind, l, kernel_choice = "Linear", num_derandom = 1)
}
\arguments{
\item{k}{Number of nearest neighbors used}

\item{X}{Observed sample X}

\item{Y}{Observed sample Y}

\item{nn_ind}{Nearest neighbor matrix of size n by k; the first column is the i-th unit value}

\item{l}{The hyperparameter for Gaussian kernel}

\item{kernel_choice}{Choice of kernel function, the default is linear kernel}

\item{num_derandom}{Number of derandomized sample used}
}
\value{
Test statistic
}
\description{
Define test statistic
}
